+++ docs/articles/release_notes.md
@@ -3,6 +3,7 @@
 * Some calls to GetCellCorner are now more efficient
 * Some fixes for CompoundGrid.
 * Added [ModHexGrid](xref:Sylves.ModHexGrid)
+* [Fix](https://github.com/BorisTheBrave/sylves/issues/6) GetChildTriangles for pointy topped hex grids.
 
 # v1.1.0
 * Added [`GetAabb`](xref:Sylves.IGrid.GetAabb(Sylves.Cell)) methods